Updates to PWR and RCC to support PLL configuration.
PLL configuration on the H7 is pretty involved, and takes a number of
configurations to make it work. In order to make peripheral drivers a bit
easier to implement, working with a soft clock tree in the rcc module which
stores the clock settings for each clock as they are setup such that users
can request the clock value from the RCC module for configuration.

Clock configuration is still missing values for setting up all of the kernel
clocks for the peripherals, but this is in work.

Added LDO settings to allow setting the scaling to support high frequencies.