Add additional definitions to STM32 FMC
Looking at the H7, F4, and F3; it appears that MWID and MTYP are all
common to each other so from this limited sample I'll put the superset
of options into the common header.

The H7 also has bank mapping options that are unique to it, so add those
into the H7 specific header.