STM32H7 ADC: Correctly return clock scaling value and fix typo
The shifting magic used before neglected to account for the fact that
the scaling sets are not powers of 2. Correct that by using actual
tables. Also account for the fact that prescaling doesn't work except
if you're using a clock other than the AHB.

While we're in here, correctly name the oversampling registers.