STM32H7 ADC: Complete Register Renaming (OVSS)
There was a typo in a register name in a header that was corrected in a
previous commit; but was referenced in C files. This commit corrects the
C files as well.

Background on the OVSS and LSHIFT; OVSS shifts the data right after
oversampling. LSHIFT shifts the data left. It it not clear the order of
operations and so we should clear both registers when setting right or
left data alignment.